requirements.txt
openai>=1.0.0
gTTS>=2.5.0
python-dotenv>=1.0.0
